logging-service/requirements.txt
fastapi
uvicorn
hazelcast-python-client
httpx
